Totally rework the structure of the IR
This is the natural, elegant, obvious way of converting constant-index memory
references into register variables: explicit loads and stores. Every operation
is now done on a Temp, which will eventually get allocated to a register. I can
optimize away redundant loads and stores near-trivially within a basic block,
and probably with somewhat more effort across blocks as well.

The IR is shaping up to be SSA, although I've punted on the hard part for now.
I'll probably end up using the functional-SSA style where each block ends with
a tail-call to another block, and blocks begin with a special instruction that
captures the arguments of those tail-calls.

x86 codegen is broken by this, and I'll have to write a register allocator
before it'll work again. The existing constant propagation implementation is
deleted, since I'll need to write a new one from scratch on this new IR (it
should be more elegant and powerful).