Add load/store elimination and DCE
DCE is near-trivial in this system. Load/store elimination is also pretty
simple. Unfortunately it doesn't work on non-constant addresses right now, so
we're going to need some powerful constant propagation before it can get much
done.

This breaks the register allocator by (somehow) producing code that runs out of
registers. This is clearly baloney, but I'll fix it later.