Fix a regalloc bug and add a jump to termination
The register allocator wasn't taking into account the liveness of the temp
holding the pointer value on the way out of the block, so it was getting bashed
by one of the temps in the jump instruction.

With this, everything works, with optimizations off. With optimizations on,
mandelbrot works (not much faster than with optimizations off... ???) but quine
doesn't.