Deal with word and byte values separately
Yes, this is, for the most part, buffoonery. There's no provision for converting
one type to the other (which may happen eventually, with optimizations). There
are just these two types. The boolean "is constant" and "is word" flags seem
janky (especially "is word").

WHATEVER. EVERYTHING RUNS NOW. OPTIMIZATIONS ON AND OFF. INSANITY WOLF.